latest updates to backend

+++ app/routers/question.py
@@ -6,6 +6,7 @@ from sqlalchemy import func
 # from sqlalchemy.sql.functions import func
 from .. import models, schemas, oauth2
 from ..database import get_db
+from datetime import timedelta, datetime
 
 
 router = APIRouter(
@@ -16,10 +17,6 @@ router = APIRouter(
 @router.post("/", response_model=schemas.QuestionOut)
 def create_question(question:schemas.Question, user:schemas.TokenUser=Depends(oauth2.get_current_user),
                     db:Session=Depends(get_db)):
-    assessment = db.query(models.Assessment).filter(models.Assessment.id == question.assessment_id).first()
-    if not assessment:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail="assessment not found")
     instructor = db.query(models.Assessment).join(
         models.CourseInstructor, models.Assessment.course_id == models.CourseInstructor.course_code
     ).filter(models.Assessment.id == question.assessment_id, 
@@ -27,6 +24,14 @@ def create_question(question:schemas.Question, user:schemas.TokenUser=Depends(oa
              models.CourseInstructor.is_accepted == True).first()
     if not instructor:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    assessment = db.query(models.Assessment).filter(models.Assessment.id == question.assessment_id).first()
+    if not assessment:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail="assessment not found")
+    current_time = datetime.now()
+    if (assessment.start_date <= current_time):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail="cannot add questions to already started assessment")
     new_question = models.Question(**question.dict())
     db.add(new_question)
     db.commit()
@@ -36,10 +41,6 @@ def create_question(question:schemas.Question, user:schemas.TokenUser=Depends(oa
 @router.put("/{id}", response_model=schemas.QuestionOut)
 def update_question(id:int, updated_question:schemas.QuestionUpdate, user:schemas.TokenUser=Depends(oauth2.get_current_user),
                     db:Session=Depends(get_db)):
-    question_query = db.query(models.Question).filter(models.Question.id==id)
-    if not question_query.first():
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail="question not found")
     instructor = db.query(models.Question).join(
         models.Assessment, models.Question.assessment_id == models.Assessment.id).join(
         models.CourseInstructor, models.Assessment.course_id == models.CourseInstructor.course_code
@@ -48,6 +49,16 @@ def update_question(id:int, updated_question:schemas.QuestionUpdate, user:schema
              models.CourseInstructor.is_accepted == True).first()
     if not instructor:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    question_query = db.query(models.Question).filter(models.Question.id==id)
+    question = question_query.first()
+    if not question:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail="question not found")
+    assessment = db.query(models.Assessment).filter(models.Assessment.id == question.assessment_id).first()
+    current_time = datetime.now()
+    if (assessment.start_date <= current_time):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail="cannot add questions to already started assessment")
     question_query.update(updated_question.dict(), synchronize_session=False)
     db.commit()
     db.refresh(question_query.first())
@@ -56,10 +67,6 @@ def update_question(id:int, updated_question:schemas.QuestionUpdate, user:schema
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_question(id:int,user:schemas.TokenUser=Depends(oauth2.get_current_user),
                     db:Session=Depends(get_db)):
-    question_query = db.query(models.Question).filter(models.Question.id==id)
-    if not question_query.first():
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail="question not found")
     instructor = db.query(models.Question).join(
         models.Assessment, models.Question.assessment_id == models.Assessment.id).join(
         models.CourseInstructor, models.Assessment.course_id == models.CourseInstructor.course_code
@@ -68,6 +75,16 @@ def delete_question(id:int,user:schemas.TokenUser=Depends(oauth2.get_current_use
              models.CourseInstructor.is_accepted == True).first()
     if not instructor:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    question_query = db.query(models.Question).filter(models.Question.id==id)
+    question = question_query.first()
+    if not question:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail="question not found")
+    assessment = db.query(models.Assessment).filter(models.Assessment.id == question.assessment_id).first()
+    current_time = datetime.now()
+    if (assessment.start_date <= current_time):
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail="cannot delete questions to already started assessment")
     question_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
